Fix: Replaced Entities With Models In Repository CRUD Operations
I essentially changed the GetById() and Create() methods for each entity Repository such that they use/return models instead of entities. I left the Update() and Delete() methods unchanged, due to the fact that in both operations we require to know the exact entity for which we perform the changes, whilst for the GetAll(), GetById() and Create() methods we either return or create new entities, without changing the ones that already exist. Furthermore, I've left the GetAllQuery() methods unchanged as well, since we only use them as helper methods for our GetAll() methods. I've chosen not to push a couple of new database-input.sql files I've created, since they donot relate to this commit's subject.