Follow-up Fix: Replaced Entities With Models In Update + Delete Methods
In the end I've decided to replace all of the entities with models in all of the CRUD operations, including Update and Delete. Since it was mainly a layer breach if we kept the entity object in the Controller or BLL layers and since it was pretty much impossible to determine which exact row we're referring to only from a basic model, I've decided to add the id of the row as well in the parameters for the Update and Delete operations. I could've chosen to not allow users or anyone for that matter to create/post rows with exact same values, but this couldn't have kept them from later updating a row to match another row 1:1 and it would have taken a lot from the overall functionality/modularity/fun of the student-platform website.